Fix cucumber tests messing with tests of dependants
The cucumber @after hook was messing with the tests in projects depending
on this framework. Because the @after hook is run after every scenario
dependants had to either overwrite the file by including a file with the same name
at the same place or do something equally ugly.

By tagging the hook and the only scenario included in the framework with the
"@Framework" tag the hook is only ran for that scenario. Thus the hook is not
ran in tests in projects depending on this one while not influencing the testing in
this project.