Made put() in Square public
As students following the SKT course, we want to use the put() method in
the Square class to put special items (such as a strawberry or a melon) on
a map to enhance the functionality of the game. However, because put() was
protected we could not use it, a simple solution is to make put() public.